test(usage): split pricing provenance tests into focused file
The runtime-llm test file exceeded the 1000-line lint cap; move the
three pricing-provenance cases (unavailable / marked placeholder-zero /
confirmed zero) into runtime-llm.usage-pricing.test.ts with their own
minimal mock wiring.
