+++ src/plugins/runtime/runtime-llm.runtime.test.ts
@@ -713,171 +713,6 @@ describe("runtime.llm.complete", () => {
     expectFields(requireRecord(logPayload.usage, "log usage"), { costUsd: 0.0042 });
   });
 
-  it("omits costUsd when model pricing is unavailable", async () => {
-    hoisted.prepareSimpleCompletionModelForAgent.mockResolvedValue(
-      createPreparedModel("unpriced-model"),
-    );
-    hoisted.resolveSimpleCompletionSelectionForAgent.mockImplementation(
-      (params: { modelRef?: string; agentId: string }) => ({
-        provider: "openai",
-        modelId: "unpriced-model",
-        agentDir: `/tmp/${params.agentId}`,
-      }),
-    );
-    hoisted.completeWithPreparedSimpleCompletionModel.mockResolvedValue({
-      content: [{ type: "text", text: "done" }],
-      // Mirror the real adapter shape: usage always carries a cost object whose
-      // totals default to zero when the provider did not bill anything.
-      usage: {
-        input: 11,
-        output: 7,
-        cacheRead: 5,
-        cacheWrite: 2,
-        total: 25,
-        cost: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0, total: 0 },
-      },
-    });
-
-    const llm = createRuntimeLlm({
-      getConfig: () => cfg,
-      authority: { allowComplete: true },
-    });
-    const result = await llm.complete({
-      messages: [{ role: "user", content: "Ping" }],
-      purpose: "test-purpose",
-    });
-
-    const usage = requireRecord(result.usage, "completion usage");
-    expect(usage.costUsd).toBeUndefined();
-    expectFields(usage, {
-      inputTokens: 11,
-      outputTokens: 7,
-      cacheReadTokens: 5,
-      cacheWriteTokens: 2,
-      totalTokens: 25,
-    });
-  });
-
-  it("omits costUsd for placeholder-zero pricing marked pricingUnavailable", async () => {
-    const cfgWithUnknownPricing = {
-      ...cfg,
-      models: {
-        providers: {
-          codex: {
-            baseUrl: "https://chatgpt.com/backend-api",
-            models: [
-              {
-                id: "gpt-unknown",
-                name: "gpt-unknown",
-                reasoning: false,
-                input: ["text"],
-                cost: {
-                  input: 0,
-                  output: 0,
-                  cacheRead: 0,
-                  cacheWrite: 0,
-                  pricingUnavailable: true,
-                },
-                contextWindow: 128_000,
-                maxTokens: 4096,
-              },
-            ],
-          },
-        },
-      },
-    } satisfies OpenClawConfig;
-    const unknownPrepared = createPreparedModel("gpt-unknown");
-    hoisted.prepareSimpleCompletionModelForAgent.mockResolvedValue({
-      ...unknownPrepared,
-      selection: { ...unknownPrepared.selection, provider: "codex" },
-    });
-    hoisted.resolveSimpleCompletionSelectionForAgent.mockImplementation(
-      (params: { modelRef?: string; agentId: string }) => ({
-        provider: "codex",
-        modelId: "gpt-unknown",
-        agentDir: `/tmp/${params.agentId}`,
-      }),
-    );
-    hoisted.completeWithPreparedSimpleCompletionModel.mockResolvedValue({
-      content: [{ type: "text", text: "done" }],
-      usage: {
-        input: 11,
-        output: 7,
-        cacheRead: 5,
-        cacheWrite: 2,
-        total: 25,
-        cost: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0, total: 0 },
-      },
-    });
-
-    const llm = createRuntimeLlm({
-      getConfig: () => cfgWithUnknownPricing,
-      authority: { allowComplete: true },
-    });
-    const result = await llm.complete({
-      messages: [{ role: "user", content: "Ping" }],
-      purpose: "test-purpose",
-    });
-
-    expect(requireRecord(result.usage, "completion usage").costUsd).toBeUndefined();
-  });
-
-  it("keeps costUsd 0 for confirmed zero-priced models", async () => {
-    const cfgWithFreeModel = {
-      ...cfg,
-      models: {
-        providers: {
-          openai: {
-            baseUrl: "https://api.openai.com/v1",
-            models: [
-              {
-                id: "free-model",
-                name: "free-model",
-                reasoning: false,
-                input: ["text"],
-                cost: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0 },
-                contextWindow: 128_000,
-                maxTokens: 4096,
-              },
-            ],
-          },
-        },
-      },
-    } satisfies OpenClawConfig;
-    hoisted.prepareSimpleCompletionModelForAgent.mockResolvedValue(
-      createPreparedModel("free-model"),
-    );
-    hoisted.resolveSimpleCompletionSelectionForAgent.mockImplementation(
-      (params: { modelRef?: string; agentId: string }) => ({
-        provider: "openai",
-        modelId: "free-model",
-        agentDir: `/tmp/${params.agentId}`,
-      }),
-    );
-    hoisted.completeWithPreparedSimpleCompletionModel.mockResolvedValue({
-      content: [{ type: "text", text: "done" }],
-      usage: {
-        input: 11,
-        output: 7,
-        cacheRead: 5,
-        cacheWrite: 2,
-        total: 25,
-        cost: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0, total: 0 },
-      },
-    });
-
-    const llm = createRuntimeLlm({
-      getConfig: () => cfgWithFreeModel,
-      authority: { allowComplete: true },
-    });
-    const result = await llm.complete({
-      messages: [{ role: "user", content: "Ping" }],
-      purpose: "test-purpose",
-    });
-
-    expect(requireRecord(result.usage, "completion usage").costUsd).toBe(0);
-  });
-
   it("preserves the completion options shape when reasoning is omitted", async () => {
     const llm = createRuntimeLlm({
       getConfig: () => cfg,
